// src/app/models/InternshipData.ts
export interface InternshipData {
  BINUSIAN_ID: string;
  STUDENT_ID: string;
  STUDENT_NAME: string;
  ACADEMIC_PROGRAM_ID: string;
  ACADEMIC_GROUP: string;
  ACADEMIC_PROGRAM: string;
  STUDENT_EMAIL: string;
  STUDENT_PHONE: string;
  CAMPUS: string;
  TRACK_ID: string;
  TRACK: string;
  GPA: string;
  TOTAL_SKS: string;
  STATUS: string;
  PARTNER_LECTURER: string;
  POSITION_TOPIC: string;
  DURATION: string;
  JOB_START_DATE: string;
  JOB_END_DATE: string;
  WORK_SCHEMA: string;
  ENROLLMENT: string;
}